Moving get_connection() from datastore.__init__ to _implicit_environ.
